RAG-deploy/4_embedding_store/requirements.txt
torch
numpy
supabase
psycopg2
transformers
python-dotenv
fastapi
uvicorn
python-multipart
# pip uninstall torch torchvision torchaudio
# then do this:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121